+++ apps/mcp-server/RELEASES.md
@@ -23,3 +23,11 @@ Based on:
 - Speakeasy CLI 1.770.0 (2.893.0) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [mcp-typescript v0.0.3] apps/mcp-server
+
+## 2026-06-05 15:49:19
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.771.0 (2.893.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [mcp-typescript v0.1.0] apps/mcp-server